+++ src/engines/Cesium/hooks/useMount.ts
@@ -1,0 +1,25 @@
+import { useCallback } from "react";
+
+export default ({
+  onMount,
+  mountCamera,
+  unmountCamera,
+}: {
+  onMount?: () => void;
+  mountCamera?: () => void;
+  unmountCamera?: () => void;
+}) => {
+  const handleMount = useCallback(() => {
+    mountCamera?.();
+    onMount?.();
+  }, [mountCamera, onMount]);
+
+  const handleUnmount = useCallback(() => {
+    unmountCamera?.();
+  }, [unmountCamera]);
+
+  return {
+    handleMount,
+    handleUnmount,
+  };
+};
